migration-tests/requirements.txt
# Dependências do migration-tests
# pip install -r requirements.txt

# PySpark + Delta (alinhe com o DBR do cluster)
pyspark>=3.4.0,<3.6.0
delta-spark>=2.4.0,<3.1.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
chispa>=0.9.4

# Utilitários
faker==24.0.0
python-dateutil>=2.8.2
